requirements.txt
flask_sqlalchemy==3.1.1
bcrypt==4.2.0
flasgger==0.9.7.1
alembic==1.13.1
Flask-Migrate==4.0.7
psycopg2==2.9.9